requirements/requirements-base.txt
# django, rest-framework
# =====================
Django>=3.2.13,<4
django-filter==21.1
djangorestframework==3.13.1
django-cors-headers==3.12.0
drf-spectacular==0.22.1
python-dotenv==0.20.0
django-model-utils==4.2.0
